Perbaikan waktu dan undo

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('alamat')->nullable();
 
             $table->string('email')->unique();
-            $table->integer('telepon')->nullable();
+            $table->bigInteger('telepon')->nullable();
             $table->boolean('status_info');
             $table->string('password');
             $table->rememberToken();
